+++ src/solosequences.rs
@@ -8,27 +8,27 @@ use crate::music_constants::*;
 pub fn get_solosequence(ss_title: &str) -> SoloSequence {
     let mut solosequences: Vec<SoloSequence> = vec![
         SoloSequence {
-                title: "FunctionsTEST",
-                instrument: 74, //74 = RECORDER
-                velocity: 100,
-                abstract_notes: vec![
-                    // REST, UNISON, UNISON, UNISON, UNISON,MAJ_2, MAJ_2, MAJ_2, MAJ_2, MAJ_2, 
-                    // MIN_3, MIN_3,MIN_3,MIN_3,MIN_3,MAJ_3, P_4, P_4, P_4, P_4, AUM_4, P_5, P_5, P_5, P_5, MAJ_6, MAJ_6, MAJ_6, MAJ_6, MIN_7, MIN_7, MIN_7, MIN_7, MIN_7, MAJ_7
-                    //the same scale in halftone numbers:
-                    -1, 0, 0, 0, 0, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 
-                    4, 5, 5, 5, 5, 6, 7, 7, 7, 7, 
-                    9, 9, 9, 9, 10, 10, 10, 10, 10, 11,
-                ],
-                octaves: vec![2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3, 4, 4],
-                figures: vec![1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 4, 5, 6],
-                iter: 200,
-                interval_time: SIXTEENTH as u64, // 60 = 1/32
-                bpm: 108,
-                check_n_replace: vec![
-                    (trill_on_interval(vec![MAJ_2, SIXTEENTH], vec![SIXTEENTH])),
-                ],
-            },
-            SoloSequence {
+            title: "FunctionsTEST",
+            instrument: 74, //74 = RECORDER
+            velocity: 100,
+            abstract_notes: vec![
+                // REST, UNISON, UNISON, UNISON, UNISON,MAJ_2, MAJ_2, MAJ_2, MAJ_2, MAJ_2, 
+                // MIN_3, MIN_3,MIN_3,MIN_3,MIN_3,MAJ_3, P_4, P_4, P_4, P_4, AUM_4, P_5, P_5, P_5, P_5, MAJ_6, MAJ_6, MAJ_6, MAJ_6, MIN_7, MIN_7, MIN_7, MIN_7, MIN_7, MAJ_7
+                //the same scale in halftone numbers:
+                -1, 0, 0, 0, 0, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 
+                4, 5, 5, 5, 5, 6, 7, 7, 7, 7, 
+                9, 9, 9, 9, 10, 10, 10, 10, 10, 11,
+            ],
+            octaves: vec![2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3, 4, 4],
+            figures: vec![1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 4, 5, 6],
+            iter: 200,
+            interval_time: SIXTEENTH as u32, // 60 = 1/32
+            bpm: 108,
+            check_n_replace: vec![
+                (trill_on_interval(vec![MAJ_2, SIXTEENTH], vec![SIXTEENTH])),
+            ],
+        },
+        SoloSequence {
             title: "PastoralSolo",
             instrument: 74, //74 = RECORDER
             velocity: 100,
@@ -43,7 +43,7 @@ pub fn get_solosequence(ss_title: &str) -> SoloSequence {
             octaves: vec![2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3, 4, 4],
             figures: vec![1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 4, 5, 6],
             iter: 200,
-            interval_time: SIXTEENTH as u64, // 60 = 1/32
+            interval_time: SIXTEENTH as u32, // 60 = 1/32
             bpm: 108,
             check_n_replace: vec![
                 (trill_on_interval(vec![MAJ_2, SIXTEENTH], vec![SIXTEENTH])),
@@ -70,7 +70,7 @@ pub fn get_solosequence(ss_title: &str) -> SoloSequence {
                 (gracenote_on_interval(vec![UNISON,SIXTEENTH], vec![-MAJ_2, SIXTY_FOURTH / 2])),
                 (gracenote_on_interval(vec![UNISON,SIXTEENTH], vec![-MIN_3, SIXTY_FOURTH / 2])),
             ],
-        },
+    },
         SoloSequence {
             title: "TheAloneLocrio",
             instrument: 75, //74 = FLUTE
@@ -84,7 +84,7 @@ pub fn get_solosequence(ss_title: &str) -> SoloSequence {
             ],
             figures: vec![1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 4, 5, 6, 7],
             iter: 500,
-            interval_time: EIGHTH as u64, // 60 = 1/32
+            interval_time: EIGHTH as u32, // 60 = 1/32
             bpm: 90,
             check_n_replace: vec![
                 (
@@ -110,7 +110,7 @@ pub fn get_solosequence(ss_title: &str) -> SoloSequence {
                 //     (
                 //         // check function
                 //         Box::new(
-                //             |args: Vec<i64>, a_pitch: i8, _a_dur: u64, b_pitch: i8, _b_dur: u64| {
+                //             |args: Vec<i32>, a_pitch: i8, _a_dur: u32, b_pitch: i8, _b_dur: u32| {
                 //                 if a_pitch == -1 || b_pitch == -1 {
                 //                     return false;
                 //                 }
@@ -122,7 +122,7 @@ pub fn get_solosequence(ss_title: &str) -> SoloSequence {
                 //     (
                 //         // replace function
                 //         Box::new(
-                //             |_args: Vec<i64>, _a_pitch: i8, _a_dur: u64, _b_pitch: i8, _b_dur: u64| {
+                //             |_args: Vec<i32>, _a_pitch: i8, _a_dur: u32, _b_pitch: i8, _b_dur: u32| {
                 //                 (vec![], vec![])
                 //             },
                 //         ),
@@ -159,14 +159,14 @@ pub struct SoloSequence<'a> {
     pub octaves: Vec<i32>,
     pub figures: Vec<u32>,
     pub iter: u32,
-    pub interval_time: u64,
+    pub interval_time: u32,
     pub bpm: u32,
     pub check_n_replace: Vec<(
         // a vector filled by tuples with (a (check closure + args) and a (replace closure + args) )
-        (Box<dyn Fn(Vec<i64>, i8, u64, i8, u64) -> bool>, Vec<i64>),
+        (Box<dyn Fn(Vec<i32>, i8, u32, i8, u32) -> bool>, Vec<i32>),
         (
-            Box<dyn Fn(Vec<i64>, i8, u64, i8, u64) -> (Vec<i8>, Vec<u64>)>,
-            Vec<i64>,
+            Box<dyn Fn(Vec<i32>, i8, u32, i8, u32) -> (Vec<i8>, Vec<u32>)>,
+            Vec<i32>,
         ),
     )>,
 }
